fix kern table error

+++ truetype/truetype.go
@@ -345,7 +345,7 @@ func (f *Font) parseKern() error {
 		return UnsupportedError(fmt.Sprintf("kern coverage: 0x%04x", coverage))
 	}
 	f.nKern, offset = int(u16(f.kern, offset)), offset+2
-	if 6*f.nKern != length-14 {
+	if uint16(6*f.nKern) != uint16(length-14) {
 		return FormatError("bad kern table length")
 	}
 	return nil
